fix:접근제어 오류 수정, 기능 구현 완료
