+++ frontend/src/app/(after-login)/posts/new/page.tsx
@@ -1,30 +1,11 @@
 'use client'
 
 import { PostContentForm } from '@/components/post-create'
-import { useRouter } from 'next/navigation'
-import { parseCookies } from 'nookies'
 import Modal from '@/components/modal-accesscontrol'
-import { useEffect, useState } from 'react'
+import { useAuth } from '@/app/_api/useAuth'
 
 export default function PostCreateForm() {
-    const [showModal, setShowModal] = useState(false)
-    const router = useRouter()
-
-    useEffect(() => {
-        const cookies = parseCookies()
-        // const token = cookies.token
-        // console.log('Parsed Cookies:', token) // 여기에서 쿠키가 올바르게 파싱되는지 확인
-
-        if (!cookies) {
-            setShowModal(true)
-            return
-        }
-    }, [router])
-
-    const handleCloseModal = () => {
-        setShowModal(false)
-        router.push('/posts')
-    }
+    const { showModal, handleCloseModal } = useAuth() // useAuth 훅을 사용하여 모달 제어
 
     return (
         <div className='space-y-6 p-10 pb-16 md:block'>
